+++ baguni-batch/src/main/java/baguni/batch/config/OllamaApiConfig.java
@@ -12,7 +12,7 @@ import org.springframework.web.client.RestClient;
 import org.springframework.web.client.support.RestClientAdapter;
 import org.springframework.web.service.invoker.HttpServiceProxyFactory;
 
-import baguni.batch.domain.feed.util.OllamaApi;
+import baguni.batch.domain.analyzer.ollama.OllamaApi;
 
 @Configuration
 public class OllamaApiConfig {
